packages/destination-actions/src/destinations/loops/createOrUpdateContact/__tests__/index.test.ts
import nock from 'nock'
import { createTestIntegration } from '@segment/actions-core'
import Destination from '../../index'

const testDestination = createTestIntegration(Destination)

const LOOPS_API_KEY = 'some random secret'

describe('Loops.createOrUpdateContact', () => {
  it('should validate action fields', async () => {
    try {
      await testDestination.testAction('createOrUpdateContact', {
        settings: { apiKey: LOOPS_API_KEY }
      })
    } catch (err) {
      expect(err.message).toContain("missing the required field 'email'.")
      expect(err.message).toContain("missing the required field 'userId'.")
    }
  })

  it('should work', async () => {
    const testPayload = {
      email: 'test@example.com',
      firstName: 'Ellen',
      lastName: 'Richards',
      source: 'Segment',
      userGroup: 'Alum',
      userId: 'some-id-1'
    }
    nock('https://app.loops.so/api/v1').put('/contacts/update', testPayload).reply(200, {
      success: true,
      id: 'someId'
    })

    const responses = await testDestination.testAction('createOrUpdateContact', {
      mapping: testPayload,
      settings: { apiKey: LOOPS_API_KEY }
    })

    expect(responses.length).toBe(1)
    expect(responses[0].status).toBe(200)
  })

  it('should work without optional fields', async () => {
    const testPayload = {
      email: 'test@example.com',
      userId: 'some-id-1'
    }
    nock('https://app.loops.so/api/v1').put('/contacts/update', testPayload).reply(200, {
      success: true,
      id: 'someId'
    })

    const responses = await testDestination.testAction('createOrUpdateContact', {
      mapping: testPayload,
      settings: { apiKey: LOOPS_API_KEY }
    })

    expect(responses.length).toBe(1)
    expect(responses[0].status).toBe(200)
  })
})
